Changed ItemQuantityFether to be a non-inner class.
